Project 4 - Dockerising multiple applications/rss_reader/requirements.txt
feedparser==6.0.10
requests==2.28.2
psycopg2-binary==2.9.5
